initial port of unpack/identify CLI implementation by Savino
